(chore) move NewItemTemplate into logbooks/
