+++ client/src/components/logbooks/NewItemTemplate/MiniField.tsx
