+++ client/src/components/logbooks/NewItemTemplate/style/NewItemTemplate.style.ts
